refactor: api 스펙 변경 및 리팩토링

+++ src/main/java/com/hhplus/concertreservation/concert/presentation/dto/response/ConcertSeatResponse.java
@@ -8,6 +8,6 @@ public record ConcertSeatResponse(
         Long seatId,
         String number,
         ConcertSeatStatus status,
-        BigDecimal price
+        int price
 ) {
 }
