+++ src/main/java/com/hhplus/concertreservation/user/presentation/dto/request/ChargeUserPointRequest.java
@@ -1,6 +1,6 @@
 package com.hhplus.concertreservation.user.presentation.dto.request;
 
 public record ChargeUserPointRequest(
-        long amount
+        int amount
 ) {
 }
