+++ src/main/java/com/hhplus/concertreservation/user/presentation/dto/response/ChargeUserPointResponse.java
@@ -1,6 +1,6 @@
 package com.hhplus.concertreservation.user.presentation.dto.response;
 
 public record ChargeUserPointResponse(
-        long totalAmount
+        int totalAmount
 ) {
 }
