+++ src/main/java/com/hhplus/concertreservation/user/presentation/dto/response/GetUserPointResponse.java
@@ -1,6 +1,6 @@
 package com.hhplus.concertreservation.user.presentation.dto.response;
 
 public record GetUserPointResponse(
-        long totalAmount
+        int totalAmount
 ) {
 }
